402.fhe/agent/requirements.txt
requests
eth-account
web3
python-dotenv
